linda-server/requirements.txt
fastapi[standard]
strawberry-graphql
repository_sqlalchemy
alembic==1.13.3
psycopg2-binary~=2.9.5
PyJWT
bcrypt
python-dotenv
pydantic-settings